Start tidying up the traits a bit
